result_2/comments/requirements.txt
asgiref==3.2.3
Django==3.0.2
django-clear-cache==0.3
django-cors-headers==3.2.1
django-static-jquery==2.1.4
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.0
